# shrink{h,v}: improve performance