- Commit 4395dd0a098ac626bef5b8f1673272e18fa3010c is a revert of commit 9b968491d366a68ea1008ed29a74f7293501e4c4. See https://gist.github.com/kleisauke/dcf24beb95cdf2b3037ff63f71b7f9bd for the rationale behind this.
- Commit e1698911402a2f3caac7131e37a8e438e1ab0827 takes a similiar approach as PR #3052 in `shrinkh`.
- Commit 77e8451f13c0aeb949648b299a4148eaec2584a9 removes the `shrinkv` sum buffer in favor of a local, ~7% faster. In line with commit 2e5880b69cf79715ac84abb28b8482aefe7df532.
- Commit 3d28a9222fcf8cd4b49f5b20e716096e60c22b38 uses `gint64` as accumulator type for the uint/int paths. In line with PR #3514.
- Commit 2d8f1c6a1b31e3cb6aef2282f83ce0f66d66778a adds a fixed-point arithmetic path for uchar images, ~9% faster.

<details>
  <summary>Test script</summary>

Test image: https://github.com/kleisauke/vips-microbench/blob/master/images/x.jpg
```python
#!/usr/bin/env python

import sys
import pyvips

im = pyvips.Image.new_from_file(sys.argv[1])

for fac in [2, 4, 8, 16]:
    for fmt in ["uchar", "ushort", "uint",
                "char", "short", "int",
                "float", "double",
                "complex", "dpcomplex"]:
        x = im.cast(fmt, shift=True)
        s = x.shrink(fac, fac)
        d = abs(s.avg() - x.avg())
        print(d, fmt)
        assert d < 2
```
</details>