@jcupitt 2024-11-21
You could keep the mul + shift for a fast uint8 path I guess. You could do uint16 and uint32 as well, but that's such a rare case it's probably not worth it.

This is a very unintuitive speedup for me! I was sure the accumulator array would be faster, since it would vectorise so well and would fetch pixel values in strict memory order. There must be something important I don't understand about cache lines :(