@jcupitt 2024-11-21
I built master with:

```
meson setup build --prefix ~/vips --libdir lib --optimization 3
```

ie. enable the autovectoriser, then with a 10k x 10k pixel RGB JPG I ran some tests like this:

```
$ time vips shrinkv wtc.jpg x.v 128

real	0m0.400s
user	0m0.326s
sys	0m0.082s
```

then fastest of five runs.

branch | vshrink | user time (s)
------ | ------- | -------------
master | 2       | 1.01
master | 10      | 0.45
master | 100     | 0.33
this   | 2       | 0.94
this   | 10      | 0.43
this   | 100     | 0.43

So I suppose I'd interpret that as being: 

- somewhere under 100 scanlines, the whole block being shrunk fits in L1 cache and the non-local fetches that the vertical accumulator loop performs fit, meaning non-locality doesn't matter

- at 100 and over, we're seeing some cache thrashing and the new vertical loop is a little slower

- the autovectoriser on the horizontal loop along the scanline seems to have no effect on runtime, which is strange (to me)