@kleisauke 2024-11-27
Hmm, ASan seems to be quite happy without that `vips_embed()` change. I wonder why it complained on this changeset though:
https://github.com/kleisauke/libvips/compare/optimize-shrink...shrink-highway
https://github.com/kleisauke/libvips/actions/runs/12050557645/job/33599676999

Anyways, I think the updated revision of this PR should fix the performance degradation for large `vshrink`'s.