@kmartinez 2024-11-28
Oh wow have we come full circle? John and I spent ages optimising older vips code to avoid multiply and especially divide (as they were so slow!) using shifts, look-up tables etc. With today's fast cores its easy to forget it still takes > 10 clocks to do a divide ;-)